+++ Client/mods/deathmatch/logic/CScriptDebugging.cpp
@@ -123,7 +123,7 @@ void CScriptDebugging::UpdateLogOutput()
     while (m_DuplicateLineFilter.PopOutputLine(line))
     {
         // Log it to the file if enough level
-        bool sufficientDebugLevel = CheckForSufficientDebugLevel(m_uiLogFileLevel, line.uiMinimumDebugLevel);
+        bool sufficientDebugLevel = CheckForSufficientDebugLevel(static_cast<std::uint8_t>(m_uiLogFileLevel), static_cast<std::uint8_t>(line.uiMinimumDebugLevel));
 
         if (sufficientDebugLevel)
         {
