+++ Server/mods/deathmatch/logic/CMainConfig.cpp
@@ -1,13 +1,14 @@
 /*****************************************************************************
- *
- *  PROJECT:     Multi Theft Auto v1.0
- *  LICENSE:     See LICENSE in the top level directory
- *  FILE:        mods/deathmatch/logic/CMainConfig.cpp
- *  PURPOSE:     XML-based main configuration file parser class
- *
- *  Multi Theft Auto is available from https://www.multitheftauto.com/
- *
- *****************************************************************************/
+*
+*  PROJECT:     Multi Theft Auto v1.0
+*  LICENSE:     See LICENSE in the top level directory
+*
+*  FILE:        mods/deathmatch/logic/CMainConfig.cpp
+*  PURPOSE:     XML-based main configuration file parser class
+*
+*  Multi Theft Auto is available from https://www.multitheftauto.com/
+*
+*****************************************************************************/
 
 #include "StdInc.h"
 #include "CMainConfig.h"
@@ -21,6 +22,7 @@
 #include "CConsoleCommands.h"
 #include "CHTTPD.h"
 #include "CStaticFunctionDefinitions.h"
+#include "CLanBroadcast.h"
 
 #define MTA_SERVER_CONF_TEMPLATE "mtaserver.conf.template"
 
@@ -1333,7 +1335,7 @@ bool CMainConfig::SetSetting(const SString& strName, const SString& strValue, bo
     }
     else if (strName == "fpslimit")
     {
-        return CStaticFunctionDefinitions::SetFPSLimit(atoi(strValue), bSave);
+        return CStaticFunctionDefinitions::SetFPSLimit(static_cast<unsigned short>(atoi(strValue)), bSave);
     }
     else if (strName == "networkencryption")
     {
