+++ Server/mods/deathmatch/logic/net/CSimVehiclePuresyncPacket.cpp
@@ -66,7 +66,7 @@ bool CSimVehiclePuresyncPacket::Read(NetBitStreamInterface& BitStream)
             return false;
         m_Cache.PlrPosition = position.data.vecPosition;
 
-        if (CVehicleManager::GetVehicleType(m_Cache.iModelID) == VEHICLE_TRAIN)
+        if (CVehicleManager::GetVehicleType(static_cast<unsigned short>(m_Cache.iModelID)) == VEHICLE_TRAIN)
         {
             // Train specific data
             float fRailPosition = 0.0f;
@@ -456,7 +456,7 @@ void CSimVehiclePuresyncPacket::ReadVehicleSpecific(NetBitStreamInterface& BitSt
     }
 
     // Door angles.
-    if (CVehicleManager::HasDoors(m_Cache.iModelID))
+    if (CVehicleManager::HasDoors(static_cast<unsigned short>(m_Cache.iModelID)))
     {
         SDoorOpenRatioSync door;
 
@@ -487,7 +487,7 @@ void CSimVehiclePuresyncPacket::WriteVehicleSpecific(NetBitStreamInterface& BitS
     }
 
     // Door angles.
-    if (CVehicleManager::HasDoors(m_Cache.iModelID))
+    if (CVehicleManager::HasDoors(static_cast<unsigned short>(m_Cache.iModelID)))
     {
         SDoorOpenRatioSync door;
         for (unsigned int i = 2; i < 6; ++i)
